services/auth-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eve-auth-service"
version = "1.0.0"
description = "EVE Co-Pilot Auth Service - EVE SSO Authentication"
requires-python = ">=3.11"
dependencies = [
    "eve-shared",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
